pyproject.toml
[project]
name = "qr-generator"
version = "0.1.0"
description = "A CLI tool to encode and decode QR code"
license = "MIT"
authors = [
    { name = "Nathan" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7", # Delivery Layer
    "segno>=1.6.1", # Infra Layer: Encoder
    "Pillow>=10.2.0", # Infra Layer: Image handling
    "pyzbar>=0.1.9", # Infra Layer: Decoder
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qr_tool"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]


[project.scripts]
qr_generator = "qr_tool.cli.main:run"
